requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
cffi==1.11.5
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
docopt==0.6.2
enum34==1.1.6
gitignore==0.0.8
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipaddress==1.0.19
lxml==4.2.1
parsel==1.4.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.5.0
PySocks==1.6.8
pywin32==223
queuelib==1.5.0
Scrapy==1.5.0
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
w3lib==1.19.0
win-inet-pton==1.0.1
zope.interface==4.4.3
